requirements.txt
django
djangorestframework
gunicorn
whitenoise
dj-database-url
psycopg2-binary
django-environ
django-safedelete
django-cas-ng
drf-jwt
